# Conflicts:
#	API_FrameWork/API_FrameWork.vcxproj.filters

+++ API_FrameWork/Cbullets.h
@@ -22,7 +22,7 @@ struct tagBullet
 	bool fire;
 	int count;
 	int rendercount;
-	int plussize;
+	float plussize;
 	bool isPlayerBullet;
 	bool iscollison;
 	int alpha;
